Move all commandline utils into a submodule
