+++ katprep/cli/__init__.py
@@ -1,0 +1,3 @@
+"""
+Modules for commandline functionality.
+"""
